+++ src/containerapp/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+Upcoming
++++++++
+* 'az containerapp up': fix --location comparison logic
+* 'az containerapp update': change --max-replicas limit
+* Add CLI support for containerapp ingress sticky-sessions'
+* Change quickstart image
+
 0.3.27
 ++++++
 * 'az containerapp secret set': add support for secrets from Key Vault