+++ src/containerapp/HISTORY.rst
@@ -2,8 +2,10 @@
 
 Release History
 ===============
-Upcoming
-+++++++
+0.3.28
+++++++
+* 'az containerapp secret set': fix help typo
+* 'az containerapp secret set': add more format validation for key vault secrets
 * 'az containerapp up': fix --location comparison logic
 * 'az containerapp update': change --max-replicas limit
 * Add CLI support for containerapp ingress sticky-sessions'
@@ -16,7 +18,7 @@ Upcoming
 * 'az containerapp secret show': add support for secrets from Key Vault
 
 0.3.26
-+++++++
+++++++
 * 'az containerapp exec': fix bugs for consumption workload based environment
 * 'az containerapp env create': fix bug causing --enable-workload-profiles to require an argument